Move to integration-tests/mockito.
