+++ settings.gradle
@@ -12,3 +12,4 @@ include ":robolectric-shadows/shadows-multidex"
 include ":robolectric-shadows/shadows-play-services"
 include ":robolectric-shadows/shadows-support-v4"
 include ":integration-tests/dependency-on-stubs"
+include ":integration-tests/mockito"
